Pull integration test decorators from local bower install.
For this we use https://github.com/Backbase/filesystem-bower-resolver,
which adds a bower dependency resolver plugin that copies a package from
the local filesystem.

Bower can do this out of the box, except when it encounters a git repo,
it tries to pull from git instead. That means you have to commit every
change for it to be installed, which is no good here. See also
https://github.com/bower/bower/issues/1275.