Rename global.js to polymer-decorators.js.
I called it global.js originally because I thought we might ship a
modules version and global version in the same package. I don't think
we'll do that though; we'll have two versions instead. Since you often
see just the filename (e.g. in Chrome devtools), it's nicer for it to be
more descriptive.