Actually remove typings for Reflect altogether.
We're just using any now. We don't really want to add typings to the
global namespace. If you choose not to use the metadata API, then you
shouldn't have it defined on your window. It was also causing problems
with the gen-global-typings.js script, which just does some text
transformation and would need some special handling to preserve the new
global declaration in a non-module context.