Better handling of tagname parameter vs is class property.
An exception is now thrown if both the `tagname` parameter to
`@customElement` is provided, and the static `is` class property is set
(except in the case that the `is` property is not an own-property of the
class). Previously the `is` property would always silently win over the
`tagname` when both were set.

Fixes #41.