Add @ExportDecoratedItems where needed
The @ExportDecoratedItems annotation is processed by tsickle, and results in @export annotations being emitted on all the elements the decorators are applied to. This prevents JsCompiler from renaming or deleting these symbols.

It prevents issues like:
* templates referencing a property that got renamed;
* computed properties not finding the property they depend upon because it got renamed;
* observers and listeners being dropped because nobody calls them.

It is not a full solution for template name safety, to do so one would have to generate externs from symbols referenced in the template, but it is the only solution to the second and third issue above.

Add @ExportDecoratedItems where needed

The @ExportDecoratedItems annotation is processed by tsickle, and results in @export annotations being emitted on all the elements the decorators are applied to. This prevents JsCompiler from renaming or deleting these symbols.

It prevents issues like:
* templates referencing a property that got renamed;
* computed properties not finding the property they depend upon because it got renamed;
* observers and listeners being dropped because nobody calls them.

It is not a full solution for template name safety, to do so one would have to generate externs from symbols referenced in the template, but it is the only solution to the second and third issue above.